Working on browserify with fileManager and binSocketClient
