+++ public/js/managers.js
@@ -1,6 +1,6 @@
 var EncryptionManager = require('./encryptionManager.js');
 var socketClient = require('./socketClient');
-var binSocketClient = require('./binSocketClient');
+var BinSocketClient2 = require('./binSocketClient');
 var fileManager = require('./fileManager');
 
 // Need to make ChatManager use prototype and call new on it
