+++ public/views/postDeps.jade
@@ -52,6 +52,5 @@ script(src='#{depRoot}js/managerBundle.js')
 //script(src='#{depRoot}js/socketClient.js')
 //script(src='#{depRoot}js/binSocketClient.js')
 //script(src='#{depRoot}js/fileManager.js')
-//script(src='#{depRoot}js/fileManagerBundle.js')
 script.
   hljs.initHighlightingOnLoad();
